pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "beautifulsoup4>=4.13.5",
    "boto3>=1.40.24",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "pdfminer-six>=20250506",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
